workaround: allow 1-step P2P on older kernel
This is a temporary workaround, tested on v4.19 kernel.
As soon as linux is updated to v5.6+, this commit becomes
obsolete, and should be removed immediately.

What this workaround does is it allows to use 'onestep'
time_stamping option even for 1-step P2P mechanism.

Here's a normal scenario:
	1-step E2E mode -> time_stamping = onestep
	1-step P2P mode -> time_stamping = p2p1step

Here's what this workaround does:
	1-step E2E & P2P mode -> time_stamping = onestep

Signed-off-by: Volodymyr Bendiuga <volodymyr.bendiuga@westermo.se>