driver interface : refactor the config_mode.
Clean up the mode settings, config_mode(). Before we had 4 different state
(PTP_TC_SLAVE/MASTER, PTP_OC_SLAVE/MASTER), now we only configure OC or TC clock.

Add a new api interface_tsinfo_tx_type(), this compare the drivers timestamp
capabillities. If the driver does not support onestep it is not a Dagger platform
and in that case we shall not set the driver mode. This should fix the problem with
running virtual system and Coronet platform.

Signed-off-by: Anders Öhlander <anders.ohlander@westermo.com>