sk: Run hwts_init despite timestamping type, ignore ENOTSUP
Previously, if linuxptp was running on hardware which supports PTP
which has been configured to run some type of hardware timestamping.

linuxptp gave the hardware no opportunity to deactivate its hardware
timestamping to allow software timestamping.

So, always call hwts_init to give the driver an opportunity to
reconfigure but ignore ENOTSUP in case we don't have a hardware
timestamping capable driver.

Signed-off-by: PatrikLundgren <patrik.lundgren@outlook.com>