POC: Split the power profile into version v1-2011 and v2-2017
The versions is called power_v1 and power_v2. V1 stands for the version
IEEE C37.238-2011 and the other for IEEE C37.238-2017.

The profile V1 has always a priority tag in all the packets. The V2 skip that part.

The Master mode append two TLV's into the announce packet. The two TLV's is called
ORGANIZATION_EXTENSION_TLV and ALTERNATE_TIME_OFFSET_INDICATOR(ATOI).

The Org TLV handles time inaccuracy and the ATOI mode handles time jump.
This POC does not handles the ATOI TLV.

When the switch works in Ordinary Master mode it will append two TLV's in the
end of the announce messages.
In the Transparent mode the switch will update the inaccuaracy time with its own
time error.

Signed-off-by: aggelander <aggelander@gmail.com>