Implement management response pmc for transparent clock specific data sets
There are two data sets specific to the transparent clock standard.
The default data set and the clock port data set.

The default and clock port data sets for the transparent clock is (almost) a
sub-set of the defaultDS and portDS of the ordinary|boundary clocks.
Thus, we do not store any seperate data, but simply reuse those structs.

Signed-off-by: PatrikLundgren <patrik.lundgren@outlook.com>