revertme: Add ifdef to net_tstamp_cpy to avoid redeclarations
This, as well as the patch before, is needed because the build system
does not point out the active kernel and will therefore use the wrong
net_tstamp (hence the _cpy). Some names end up being defined twice due
to the _cpy file, and are therefore set inside #ifdef in _cpy, while
they are #ifndef in missing.h (making them always exclusive).

Signed-off-by: Casper Andersson <casper.casan@gmail.com>