Add dummy pdelay_resp_fup option
Some PHYs can only choose between two-step or full p2p1step. There is no
option for normal one-step. In this case we will let it perform
p2p1step, and just send a dummy frame with TS and correctionField
unmodified. The calculations will still give the same result.

This is the formula used for two-step:
<meanLinkDelay> = [(t4 − t1) − (responseOriginTimestamp
− requestReceiptTimestamp) − <correctedPdelayRespCorrectionField>
− correctionField of Pdelay_Resp_Follow_Up]/2

t1 and t4 are taken as normal. Request and response timestamps will be
set to 0, and the residence time is instead updated in the response
correctionField (with the use of the reserved2 field) in HW. The
correctionField of the pdelay_resp_fup should be whatever the pdelay_req
had on ingress, as it would in two-step.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>