hsr/prp: Add BMCA for Boundary Clock
IEC62439-3 specifies unique behavior for devices running PTP in an HSR
ring.

The ring ports should be paired. If upstream is attached to a
GM, the ring ports should both be MASTER. Or, if it's a redundant
connection, both ports should be PASSIVE (PASSIVE_MASTER). For a BC
that is synchronized through the ring and transmitting downstream the
ring ports should be SLAVE on one and PASSIVE_SLAVE on the other.

PASSIVE_SLAVE is a new state introduced for this, along with the new
event PSLAVE. If synchronization on the SLAVE port ceases it should
transition the PASSIVE_SLAVE to SLAVE.

Within the ring, ANNOUNCE, SYNC, and FOLLOW_UP should make a full
lap in the ring, and be eliminated when received by the other ring
port. The forwarding within the ring should be done like a TC.

This commit only adds the basic BMCA functionality to achieve the
right port states in BC mode.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>