hsr/prp: Finish P2P 1-step TC
HSR should forward in HW between the ring ports to preserve the tag
and stay within the EntryForgetTime. Specific forwarding rules for
this is expected to be set up in HW when HSR is configured. Via
interlink we must go through CPU if we want to support DATC mode.

PRP takes up everything to CPU like normal PTP operation.

Both modes transmit on port A and have it duplicated in HW. If port A
is unavailable it will use port B. It will always use the identity of
port A. To support 2-step it would require the use of
a auto-fup/dummy_fup mode, meaning that you send a Follow-up without
timestamp and the HW will fill in the timestamp from the matching Sync.

Config options introduced:
- hsr_prp_mode [none|hsr|prp]. Global option, sets the mode of operation.
- hsr_prp_port_a [0|1]. Port option, marks port as port A.
- hsr_prp_port_b [0|1]. Port option, marks port as port B.

There should only be one port A and one port B

- Removed work required for 2-step. May be re-added in the future
- E2E mode requires further work

Signed-off-by: Casper Andersson <casper.casan@gmail.com>