hsr/prp: Write path delay + ingr/egr latency to HW
Used for offloaded forwarding in the HSR ring. Ingress and egress
delay needs to be included in this register for the current
platform. It could (and for better HW it should) use separate calls to
the driver. But now it seemed redundant since the current HW using
this only has a single register for it.

A proper offloaded solution would also require many more changes so I
see no need to plan for that right now.

HAS_ETHTOOL_MEAN_PATH_DELAY allows us to build this on other
targets (e.g. on host) where this is not defined in the header
files. The other options is to include the whole linux/ethtool.h, but
that feels like a more clunky solution.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>