tc_syntonize: Add check for max_freq and run every second
Large frequency changes caused it to attempt setting higher than
what's possible for the PHC. Add a max_freq limit.

With improvements to drivers it seems fine to run syntonization every
second, and gives better performances (though it might be good to have
a servo so it's less sensitive to changes).

Signed-off-by: Casper Andersson <casper.casan@gmail.com>