bmc: hsr-prp: Make sticky and handle HSR-PRP setup properly
The Master states weren't behaving correctly for an HSR-PRP device.

Sticky states for the receiver prevents unnecessary swapping, and
allows a recently connected cable to get synchronization up
again (switching directly to a cable that was just plugged in could
result in large fluctuations since pdelay might not be calculated yet).

Signed-off-by: Casper Andersson <casper.casan@gmail.com>