hsr-prp: Add extended datasets
Not entirely sure if peerIdentity is handled correctly. Ideally, it
should be based on the DMAC from the pdelay messages, but ptp4l hides
those from the protocol level.

We also don't have any way to configure an event if peerDelayLim is
exceeded. It's just hardcoded to 100 (the limit defined by the
standard).

Also, add pre_send and post_rcv for transparent_clock datasets.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>