ptp4l/pmc: Add TCP UDS socket for faster querying of local clock
Adds a separate socket by appending "-tcp" to the name of the normal
socket (default: /var/run/ptp4l, so this becomes /var/run/ptp4l-tcp).

This is only for the local clock and is not forwarded like Management
messages. It improves the querying speed significantly and can include
multiple datasets in the same query.

Not all datasets/fields are currently implemented. As a first step I
only did what I needed.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>