red: Initial draft of HSR/PRP Boundary Clock
Woah! New attempt with new abstractions.

Hide the redundant ports behind a RED interface that acts as a single
port towards the Clock. The Clock should never know that there are two
ports, is the goal at least. In some cases there needs to be special
handling, e.g. during setup and for Management messages. But it's much
cleaner compared to the previous attempt.

Still some ways of being complete.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>