red: Implement support for HSR/PRP RedBox
====================================================================
This commit is 27 commits squashed together from development of the
feature. Some commit messages are kept below to give some indication
of the thought process. Commits that were just "fixes" were squashed
and not kept.
====================================================================

Woah! New attempt with new abstractions.

Hide the redundant ports behind a RED interface that acts as a single
port towards the Clock. The Clock should never know that there are two
ports, is the goal at least. In some cases there needs to be special
handling, e.g. during setup and for Management messages. But it's much
cleaner compared to the previous attempt.

=================================================
Implementation of IEC 62439-3, HSR/PRP redundancy
=================================================

This implementation uses the term RED (short for redundancy). The
term RED, as well as some other inspiration, was taken from Texas
Instrument, but their approach is different since they rely on
software duplication and some kernel additions. My conclusion was that
we can't easily reuse their work. So the overall structure is very
different.
https://git.ti.com/cgit/processor-sdk/linuxptp/log/?h=ti-linuxptp-v0.3-am57xx-bc

Only BC is implemented. But it can probably easily be adjusted for an
offloaded TC.

Certain modes are skipped in the implementation as they likely won't
be used in this combination:
- 2-step
- E2E
- Unicast
- gPTP

The structure looks like this

        Clock `struct clock`
       /     \
    ethI     RED `struct port`
            /   \
	  ethA ethB `struct red_port`

RED is a dummy port that doesn't correspond to any real
port/interface. It only exists as a proxy for A/B. Transmitting on RED
will use an active port (A is priorized, but if it's unavailable it
will use B) which is then supposed to be duplicated in
HW. Receiving/timer actions will propagate necessary actions
upwards. When upwards propagation is not necessary it can be handled
without causing any events. E.g. switchover should not alert the clock
other than switching PHC and pdelay.

Peer delay messages must not be duplicated by HW. They are sent
separately on both ports.

RED is almost a normal port from the perspective of the Clock. With
the exception of a few things:
- Default data set needs to report the correct number of real ports for ts2phc
- Port B has a portIdentity number that doesn't exist in any `struct
port`, which must be handled for targeted Management queries (used by ts2phc).

Otherwise, a lot of the abstraction hides behind the callbacks of
red_event/red_dispatch.

The best port between ethA and ethB is selected and RED->best is set
to that one. Which is the only thing required for BMCA by the Clock.
The state PASSIVE_SLAVE is entirely internal to RED. No modifications
for BMCA have been done.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>

red: Add extended datasets for HSR/PRP

Not entirely sure if peerIdentity is handled correctly. Ideally, it
should be based on the DMAC from the pdelay messages, but ptp4l hides
those from the protocol level.

We also don't have any way to configure an event if peerDelayLim is
exceeded. It's just hardcoded to 100 (the limit defined by the
standard).

Also, add pre_send and post_rcv for transparent_clock datasets.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>

red: Change timestamp clock type for BC

When BC is Ring Master it must be in BC mode. Else it must be in TC
mode to forward within the ring.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>

red: Add support for redundant HSR-PRP BC via PRP

If two HSR-PRP BCs are in the same ring, one will be Passive Master.
Normally, since they have identical Announce messages, it will fall
back to lowest identity/MAC. But we want the startup to not interrupt
any ongoing synchronization. This commit ensures we always enter
PASSIVE if another BC of equal quality, attached to the same GM, is
already transmitting Syncs.

Taking over as Active Master in the ring before being fully
synchronized results in high inaccuracy. It's better to let the
current Active Master remain active.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>

red: Re-arm timers when a single port goes up

This is a band-aid to another problem where it would fail to TX
timestamp immediately after a port goes up. This re-arms the timers,
making the transmit happen a little later than it would have.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>

red: Add check for PS_PASSIVE on TC and go to PS_MASTER instead

This check shouldn't be needed, but due to ingress rules mirroring to
CPU it is required. The TC ends up seeing the exact same GM on
interlink and ring.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>