red: 2nd code dump (see details)
- FAULT file descriptors for red_ports [NOT FINISHED]
- Independent portIdentity for red_ports
- Management message handling
- Port state notifications
- Swap pdelay on switchover (ported from other branch)

Slave switchover works very well on link break. Still some possibly
unnecessary state changes. Announce timeout not tested. Ring Master not
tested yet, still running with the old implementation for that.

=================================================
Implementation of IEC 62439-3, HSR/PRP redundancy
=================================================

This implementation uses the term RED (short for redundancy). The
term RED, as well as some other inspiration, was taken from Texas
Instrument, but their approach is different since they rely on
software duplication and some kernel additions. My conclusion was that
we can't easily reuse their work. So the overall structure is very
different.
https://git.ti.com/cgit/processor-sdk/linuxptp/log/?h=ti-linuxptp-v0.3-am57xx-bc

Only BC is implemented. But it can probably easily be adjusted for an
offloaded TC.

Certain modes are skipped in the implementation as they likely won't
be used in this combination:
- 2-step
- E2E
- Unicast
- gPTP

The structure looks like this

        Clock `struct clock`
       /     \
    ethI     RED `struct port`
            /   \
	  ethA ethB `struct red_port`

RED is a dummy port that doesn't correspond to any real
port/interface. It only exists as a proxy for A/B. Transmitting on RED
will use an active port (A is priorized, but if it's unavailable it
will use B) which is then supposed to be duplicated in
HW. Receiving/timer actions will propagate necessary actions
upwards. When upwards propagation is not necessary it can be handled
without causing any events. E.g. switchover should not alert the clock
other than switching PHC and pdelay.

Peer delay messages must not be duplicated by HW. They are sent
separately on both ports.

RED is almost a normal port from the perspective of the Clock. With
the exception of a few things:
- Default data set needs to report the correct number of real ports for ts2phc
- Port B has a portIdentity number that doesn't exist in any `struct
port`, which must be handled for targeted Management queries (used by ts2phc).

Otherwise, a lot of the abstraction hides behind the callbacks of
red_event/red_dispatch.

The best port between ethA and ethB is selected and RED->best is set
to that one. Which is the only thing required for BMCA by the Clock.
The state PASSIVE_SLAVE is entirely internal to RED. No modifications
for BMCA have been done. Though, it still needs to advertise the state
for Management messages for ts2phc to work.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>