clock: Fix wrong boundaryHops value for OC
For an OC, `msg_ready` will never be set since it doesn't have any
ports to forward to. This resulted in the previous decrement never
being undone, and the Management reply would have incorrect
boundaryHops.

The `startingBoundaryHops - boundaryHops` calculation in
`port_management_reply` would give a new `boundaryHops` value that is
1 larger than what it should be.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>