ptpmon: Add new application
ptpmon gathers time_status_np from multiple local slaves by
subscribing to them. The data is then sent to an external server for
storage, analyzing, and visualization. The purpose is to run multiple
free-running slaves on the same device as the GM by having the PHCs
synchronized, thus the reported offset is always the true offset.

A --relative-to flag is proposed (but not yet implemented), in case
the GM is external. Then the relative-to port is supposed to be
directly connected to the GM, while the other ports may go through
many switches.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>